backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Allen Bercero",email = "allenbercero@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13.5,<4"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "black (>=25.1.0,<26.0.0)",
    "boto3 (>=1.39.14,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "reportlab (>=4.4.3,<5.0.0)",
    "pypdf (>=5.8.0,<6.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.4.1"

[tool.poetry]
package-mode = false